Fix image icon and place the interface in the own component.
